вынес вспомогательные функции в отдельный файл
