+++ js/util.js
@@ -18,7 +18,6 @@ const getRandomFloatingPointNumber = (min, max, numberAfterPoint) =>{
   }
 };
 const getСalculatingRandomNumber = (name) => name[getRandomNumber(0, name.length - 1)];
-
 const randomArrElements = (source) => {
   //Nc длина массива должна быть в диапазоне от 1 до  source.length
   const Nc = Math.floor(getRandomNumber(1, source.length)  );
